+++ app/(auth)/reset-password/page.tsx
@@ -8,7 +8,7 @@ import Title from '@/app/components/auth/Title';
 import { getFieldMessage } from '@/app/shared/consts/errorMessages';
 import { FieldStatus } from '@/app/shared/types/FormStatus';
 import { Button } from '@/app/shared/ui/button';
-import Input from '@/app/shared/ui/input/Input';
+import Input from '@/app/shared/ui/input';
 import Label from '@/app/shared/ui/label/Label';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
