+++ app/(auth)/social-info/page.tsx
@@ -10,7 +10,7 @@ import { getFieldMessage } from '@/app/shared/consts/errorMessages';
 import { useAuthStore } from '@/app/shared/stores/authStore';
 import { FormStatus } from '@/app/shared/types/FormStatus';
 import { Button } from '@/app/shared/ui/button';
-import Input from '@/app/shared/ui/input/Input';
+import Input from '@/app/shared/ui/input';
 import Label from '@/app/shared/ui/label/Label';
 import { useRouter } from 'next/navigation';
