+++ app/components/main/modal/IncomeExpenseForm.tsx
@@ -2,7 +2,7 @@
 
 import { IncomeExpenseFormProps } from '@/app/shared/types/Calendar';
 import { Button } from '@/app/shared/ui/button';
-import Input from '@/app/shared/ui/input/Input';
+import Input from '@/app/shared/ui/input';
 import { useState } from 'react';
 
 export default function IncomeExpenseForm({ onClose }: IncomeExpenseFormProps) {
